Updated validation method to use vectors of the same type than vector size_t (unsigned long long). This is consistent and prevent data loss

+++ pyvoronoi/pyvoronoi.pyx
@@ -98,8 +98,8 @@ cdef extern from "voronoi.hpp":
         void MapEdgeIndexes()
         void MapCellIndexes()
 
-        int CountPoints()
-        int CountSegments()
+        long long CountPoints()
+        long long CountSegments()
         long long CountVertices()
         long long CountEdges()
         long long CountCells()
